+++ decoder.py
@@ -6,12 +6,12 @@ raw_bytes = input_file.read()
 print(raw_bytes)
 
 try:
-	decoded_bytes = str(raw_bytes.decode('utf-8'))
+    decoded_bytes = str(raw_bytes.decode('utf-8'))
 except Exception as e:
-	print('Error decoding file as "UTF-8"!\n')
-	print(e)
-	print('Decoding file as "ISO-8859-9"\n')
-	decoded_bytes = str(raw_bytes.decode('iso-8859-9'))
+    print('Error decoding file as "UTF-8"!\n')
+    print(e)
+    print('Decoding file as "ISO-8859-9"\n')
+    decoded_bytes = str(raw_bytes.decode('iso-8859-9'))
 
 if decoded_bytes is not None:
-	print(decoded_bytes)
+    print(decoded_bytes)
